requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-generativeai
python-dotenv
watchdog
psycopg2-binary